Create Telebot class. Wrap methods with class
